feat: agrega campo detalle en factura, edición de cantidad con decimales y popup de detalle en cuadre
- Migración: añade columna `detalle TEXT` a `facturas` verificando con PRAGMA table_info
- Modelo Factura: campo `detalle: Optional[str]`
- Modelo FacturaItem: `cantidad` cambia de int a float para soportar decimales
- factura_repo: `crear()` acepta parámetro `detalle` opcional
- facturar_view: entry de nota opcional al pie del panel de factura; botón `+` en cada ítem abre popup para editar cantidad (soporta float, Enter confirma)
- cuadre_view: columna de detalle truncado (clickeable) en filas de factura; popup `_abrir_detalle` muestra texto completo

+++ app/models/factura_item.py
@@ -12,5 +12,5 @@ class FacturaItem:
     factura_id: int
     nombre: str             # snapshot del nombre al momento de la venta
     precio_unitario: float
-    cantidad: int
+    cantidad: float
     subtotal: float
